Initial import from Inception.
https://github.com/mvanbem/inception

I moved the main crate's readme to the root, but there should be a separate
readme that isn't full of rustdoc-isms and is written more like a github readme.

The overview docs are quite stale; they use some outdated terms and may refer to
features that don't exist yet or no longer exist. But the crate does work.